+++ config.py
@@ -1,3 +1,0 @@
-DEBUG = True
-SECRET_KEY = 'hin6bab8ge25*r=x&amp;+5$0kn=-#log$pt^#@vrqjld!^2ci@g*b'
-SITE_NAME = 'Cori+'